Correct error where api keys would not show or delete.
This used to work, but for some reason now, when looping over
a set of ApiKey objects using the b-table items, the data
gets flattened and is no longer an API object, but just a
standard object with the same parameters?

Lots has changed since this was originally working so I haven't
been able to pinpoint a change that has caused this.